+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - 2023-07-09: Impl inbox, CHANGELOG, gp, collect email, +6 query
 - 2023-07-09: Retry gp, mod, scheduler, collect email, +5
 - 2023-07-10: Metrics prompts, mod, mod, CHANGELOG, +5
+- 2023-07-11: Rate mod, scheduler, collect email, smtp, +4